Change 'client' class on 'scraper' class
-> Sites that we want to scrape are not clients at all
