+++ lib/quiz_scraper.rb
@@ -1,5 +1,5 @@
 require "quiz_scraper/version"
-require "quiz_scraper/client"
+require "quiz_scraper/scraper"
 require "quiz_scraper/request_handler"
 require "quiz_scraper/request_processor"
 require "clients/pub_quizzer"
